+++ src/main/java/roomescape/dto/reservation/ReservationResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import roomescape.domain.reservation.Reservation;
 import roomescape.dto.member.MemberResponse;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
+import roomescape.dto.time.ReservationTimeResponse;
 import roomescape.dto.theme.ThemeResponse;
 
 public record ReservationResponse(
