--- src/main/java/roomescape/dto/reservation/ReservationResponseWithPayment.java
@@ -1,31 +1,0 @@
-package roomescape.dto.reservation;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalDate;
-import roomescape.domain.payment.Payment;
-import roomescape.domain.reservation.Reservation;
-import roomescape.dto.member.MemberResponse;
-import roomescape.dto.payment.PaymentResponse;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
-import roomescape.dto.theme.ThemeResponse;
-
-public record ReservationResponseWithPayment(
-        Long id,
-        @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
-        ReservationTimeResponse time,
-        ThemeResponse theme,
-        MemberResponse member,
-        PaymentResponse payment
-) {
-
-    public static ReservationResponseWithPayment of(Reservation reservation, Payment payment) {
-        return new ReservationResponseWithPayment(
-                reservation.getId(),
-                reservation.getDate(),
-                ReservationTimeResponse.from(reservation.getTime()),
-                ThemeResponse.from(reservation.getTheme()),
-                MemberResponse.from(reservation.getMember()),
-                PaymentResponse.from(payment)
-        );
-    }
-}
