+++ src/main/java/roomescape/dto/time/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservationtime;
+package roomescape.dto.time;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
