+++ src/main/java/roomescape/web/controller/time/ReservationTimeController.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.reservationtime.ReservationTimeRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
-import roomescape.dto.reservationtime.TimeWithAvailableResponse;
+import roomescape.dto.time.ReservationTimeRequest;
+import roomescape.dto.time.ReservationTimeResponse;
+import roomescape.dto.time.TimeWithAvailableResponse;
 import roomescape.service.time.ReservationTimeDeleteService;
 import roomescape.service.time.ReservationTimeRegisterService;
 import roomescape.service.time.ReservationTimeSearchService;
