+++ src/test/java/roomescape/service/booking/time/ReservationTimeSearchServiceTest.java
@@ -15,8 +15,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.domain.time.ReservationTime;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
-import roomescape.dto.reservationtime.TimeWithAvailableResponse;
+import roomescape.dto.time.ReservationTimeResponse;
+import roomescape.dto.time.TimeWithAvailableResponse;
 import roomescape.exception.RoomEscapeException;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.service.time.ReservationTimeSearchService;
