+++ src/test/java/roomescape/service/booking/waiting/module/WaitingSearchServiceTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.dto.waiting.WaitingReservationResponse;
 import roomescape.service.ServiceBaseTest;
 
 class WaitingSearchServiceTest extends ServiceBaseTest {
